requirements.txt
requests==2.28.0
pandas==1.3.1
python-dotenv==0.20.0